[fix][cci] Reduct epheremal-storage request to 1024Mi.
Because kubelet calculate node allocatable epheremal-storage base on
root dir, but we use /var/lib/containerd/ to calculate the rootfs quota,
so the epheremal-storage request need to reduce to match the bug of
kubelet allocatable epheremal-storage calculation.

After we fix kubelet bug, then we can increase the request to the same
with limit.

Signed-off-by: Ketor <d.ketor@gmail.com>